// src/api/pojo/UserAddressGetAddress.ts
export class UserAddressGetAddress {
    list?: {
        id: number

        user_id: number

        receive: string

        phone: string

        province_id: number

        province: string

        city_id: number

        city: string

        address: string

        is_del: number

        is_default: number

        create_time: string
    }
}